+++ CHANGELOG.md
@@ -7,6 +7,7 @@ No other features planned. To do some fancy stuff with dictionaries, specialized
 
 - All the secondary dict-related nodes are moved to a subcategory.
 - Default node name shortened: `Format-Dict` → just `Dict`.
+- New node: `Extract String from Dict` - to be actually self-sufficient for the common use cases.
 
 # v1.0.3